api/requirements.txt
aiohttp
aiosignal
annotated-types
anyio
astrapy
attrs
certifi
charset-normalizer
click
dataclasses-json
deprecation
dnspython
dotenv
email_validator
fastapi
fastapi-cli
filelock
frozenlist
fsspec
h11
h2
hpack
httpcore
httptools
httpx
httpx-sse
huggingface-hub
hyperframe
idna
Jinja2
joblib
jsonpatch
jsonpointer
langchain
langchain-astradb
langchain-community
langchain-core
langchain-cohere
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
mpmath
multidict
mypy-extensions
networkx
numpy
orjson
packaging
pillow
propcache
pydantic
pydantic-settings
pydantic_core
Pygments
pymongo
python-dotenv
python-multipart
PyYAML
regex
requests
requests-toolbelt
rich
rich-toolkit
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
sniffio
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tokenizers
toml
torch
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
urllib3
uuid6
uvicorn
uvloop
watchfiles
websockets
yarl
zstandard
redis
